+++ packages/core/src/utils.ts
@@ -1,16 +1,19 @@
 import { createHash } from 'crypto';
-import type { CoreMessage, LanguageModelV1 } from 'ai';
+import type { WritableStream } from 'stream/web';
+import type { CoreMessage } from 'ai';
 import jsonSchemaToZod from 'json-schema-to-zod';
 import { z } from 'zod';
 import type { MastraPrimitives } from './action';
 import type { ToolsInput } from './agent';
+import type { AnyAISpan } from './ai-tracing';
 import type { IMastraLogger } from './logger';
 import type { Mastra } from './mastra';
-import type { AiMessageType, MastraMemory } from './memory';
+import type { AiMessageType, MastraLanguageModel, MastraMemory } from './memory';
 import type { RuntimeContext } from './runtime-context';
-import type { ChunkType } from './stream/MastraAgentStream';
-import type { CoreTool, ToolAction, VercelTool } from './tools';
+import type { ChunkType } from './stream/types';
+import type { CoreTool, VercelTool, VercelToolV5 } from './tools';
 import { CoreToolBuilder } from './tools/tool-builder/builder';
+import type { ToolToConvert } from './tools/tool-builder/builder';
 import { isVercelTool } from './tools/toolchecks';
 
 export const delay = (ms: number) => new Promise(resolve => setTimeout(resolve, ms));
@@ -223,11 +226,10 @@ export interface ToolOptions {
   runtimeContext: RuntimeContext;
   memory?: MastraMemory;
   agentName?: string;
-  model?: LanguageModelV1;
+  model?: MastraLanguageModel;
   writableStream?: WritableStream<ChunkType>;
+  agentAISpan?: AnyAISpan;
 }
-
-type ToolToConvert = VercelTool | ToolAction<any, any, any>;
 
 /**
  * Checks if a value is a Zod type
@@ -312,6 +314,14 @@ export function makeCoreTool(
   logType?: 'tool' | 'toolset' | 'client-tool',
 ): CoreTool {
   return new CoreToolBuilder({ originalTool, options, logType }).build();
+}
+
+export function makeCoreToolV5(
+  originalTool: ToolToConvert,
+  options: ToolOptions,
+  logType?: 'tool' | 'toolset' | 'client-tool',
+): VercelToolV5 {
+  return new CoreToolBuilder({ originalTool, options, logType }).buildV5();
 }
 
 /**