+++ packages/core/src/workflows/default.ts
@@ -1,9 +1,11 @@
 import { randomUUID } from 'crypto';
 import { context as otlpContext, trace } from '@opentelemetry/api';
 import type { Span } from '@opentelemetry/api';
+import { AISpanType, getSelectedAITracing } from '../ai-tracing';
+import type { AISpan, AnyAISpan, AITracingContext } from '../ai-tracing';
 import type { RuntimeContext } from '../di';
 import { MastraError, ErrorDomain, ErrorCategory } from '../error';
-import type { ChunkType } from '../stream/MastraAgentStream';
+import type { ChunkType } from '../stream/types';
 import { ToolStream } from '../tools/stream';
 import { EMITTER_SYMBOL } from './constants';
 import type { ExecutionGraph } from './execution-engine';
@@ -22,6 +24,7 @@ export type ExecutionContext = {
     delay: number;
   };
   executionSpan: Span;
+  aiSpan?: AISpan<AISpanType.WORKFLOW_RUN>;
 };
 
 /**
@@ -159,23 +162,58 @@ export class DefaultExecutionEngine extends ExecutionEngine {
       delay?: number;
     };
     runtimeContext: RuntimeContext;
+    parentAISpan?: AnyAISpan;
     abortController: AbortController;
     writableStream?: WritableStream<ChunkType>;
   }): Promise<TOutput> {
-    const { workflowId, runId, graph, input, resume, retryConfig } = params;
+    const { workflowId, runId, graph, input, resume, retryConfig, runtimeContext, parentAISpan } = params;
     const { attempts = 0, delay = 0 } = retryConfig ?? {};
     const steps = graph.steps;
 
     //clear runCounts
     this.runCounts.clear();
 
+    const spanArgs = {
+      name: `workflow run: '${workflowId}'`,
+      input,
+      attributes: {
+        workflowId,
+      },
+    };
+
+    // if parentSpan passed, use it to build workflowSpan
+    // otherwise, attempt to create new trace
+    let aiSpan: AISpan<AISpanType.WORKFLOW_RUN> | undefined;
+    if (parentAISpan) {
+      aiSpan = parentAISpan.createChildSpan({
+        type: AISpanType.WORKFLOW_RUN,
+        ...spanArgs,
+      });
+    } else {
+      const aiTracing = getSelectedAITracing({
+        runtimeContext: runtimeContext,
+      });
+      if (aiTracing) {
+        aiSpan = aiTracing.startSpan({
+          type: AISpanType.WORKFLOW_RUN,
+          ...spanArgs,
+          startOptions: {
+            runtimeContext,
+          },
+        });
+      }
+    }
+
     if (steps.length === 0) {
-      throw new MastraError({
+      const empty_graph_error = new MastraError({
         id: 'WORKFLOW_EXECUTE_EMPTY_GRAPH',
         text: 'Workflow must have at least one step',
         domain: ErrorDomain.MASTRA_WORKFLOW,
         category: ErrorCategory.USER,
       });
+
+      aiSpan?.error({ error: empty_graph_error });
+      throw empty_graph_error;
     }
 
     const executionSpan = this.mastra?.getTelemetry()?.tracer.startSpan(`workflow.${workflowId}.execute`, {
@@ -209,6 +247,7 @@ export class DefaultExecutionEngine extends ExecutionEngine {
             suspendedPaths: {},
             retryConfig: { attempts, delay },
             executionSpan: executionSpan as Span,
+            aiSpan,
           },
           abortController: params.abortController,
           emitter: params.emitter,
@@ -216,6 +255,7 @@ export class DefaultExecutionEngine extends ExecutionEngine {
           writableStream: params.writableStream,
         });
 
+        // if step result is not success, stop and return
         if (lastOutput.result.status !== 'success') {
           if (lastOutput.result.status === 'bailed') {
             lastOutput.result.status = 'success';
@@ -238,8 +278,26 @@ export class DefaultExecutionEngine extends ExecutionEngine {
             error: result.error,
             runtimeContext: params.runtimeContext,
           });
+
+          if (result.error) {
+            aiSpan?.error({
+              error: result.error,
+              attributes: {
+                status: result.status,
+              },
+            });
+          } else {
+            aiSpan?.end({
+              output: result.result,
+              attributes: {
+                status: result.status,
+              },
+            });
+          }
           return result;
         }
+
+        // if error occurred during step execution, stop and return
       } catch (e) {
         const error =
           e instanceof MastraError
@@ -274,10 +332,19 @@ export class DefaultExecutionEngine extends ExecutionEngine {
           error: result.error,
           runtimeContext: params.runtimeContext,
         });
+
+        aiSpan?.error({
+          error,
+          attributes: {
+            status: result.status,
+          },
+        });
+
         return result;
       }
     }
 
+    // after all steps are successful, return result
     const result = (await this.fmtReturnValue(executionSpan, params.emitter, stepResults, lastOutput.result)) as any;
     await this.persistStepUpdate({
       workflowId,
@@ -290,6 +357,14 @@ export class DefaultExecutionEngine extends ExecutionEngine {
       error: result.error,
       runtimeContext: params.runtimeContext,
     });
+
+    aiSpan?.end({
+      output: result.result,
+      attributes: {
+        status: result.status,
+      },
+    });
+
     return result;
   }
 
@@ -532,6 +607,7 @@ export class DefaultExecutionEngine extends ExecutionEngine {
     runtimeContext,
     skipEmits = false,
     writableStream,
+    serializedStepGraph,
   }: {
     workflowId: string;
     runId: string;
@@ -548,6 +624,7 @@ export class DefaultExecutionEngine extends ExecutionEngine {
     runtimeContext: RuntimeContext;
     skipEmits?: boolean;
     writableStream?: WritableStream<ChunkType>;
+    serializedStepGraph: SerializedStepFlowEntry[];
   }): Promise<StepResult<any, any, any, any>> {
     const startTime = resume?.steps[0] === step.id ? undefined : Date.now();
     const resumeTime = resume?.steps[0] === step.id ? Date.now() : undefined;
@@ -558,7 +635,17 @@ export class DefaultExecutionEngine extends ExecutionEngine {
       ...(resume?.steps[0] === step.id ? { resumePayload: resume?.resumePayload } : { payload: prevOutput }),
       ...(startTime ? { startedAt: startTime } : {}),
       ...(resumeTime ? { resumedAt: resumeTime } : {}),
+      status: 'running',
     };
+
+    const stepAISpan = executionContext.aiSpan?.createChildSpan({
+      name: `workflow step: '${step.id}'`,
+      type: AISpanType.WORKFLOW_STEP,
+      input: prevOutput,
+      attributes: {
+        stepId: step.id,
+      },
+    });
 
     if (!skipEmits) {
       await emitter.emit('watch', {
@@ -566,7 +653,6 @@ export class DefaultExecutionEngine extends ExecutionEngine {
         payload: {
           currentStep: {
             id: step.id,
-            status: 'running',
             ...stepInfo,
           },
           workflowState: {
@@ -574,7 +660,6 @@ export class DefaultExecutionEngine extends ExecutionEngine {
             steps: {
               ...stepResults,
               [step.id]: {
-                status: 'running',
                 ...stepInfo,
               },
             },
@@ -590,24 +675,46 @@ export class DefaultExecutionEngine extends ExecutionEngine {
           id: step.id,
           stepCallId,
           ...stepInfo,
-          status: 'running',
         },
       });
     }
 
+    await this.persistStepUpdate({
+      workflowId,
+      runId,
+      serializedStepGraph,
+      stepResults: {
+        ...stepResults,
+        [step.id]: stepInfo,
+      } as Record<string, StepResult<any, any, any, any>>,
+      executionContext,
+      workflowStatus: 'running',
+      runtimeContext,
+    });
+
     const _runStep = (step: Step<any, any, any, any>, spanName: string, attributes?: Record<string, string>) => {
       return async (data: any) => {
+        const aiTracingContext: AITracingContext = {
+          parentAISpan: stepAISpan,
+          metadata: {},
+        };
+
+        const enhancedData = {
+          ...data,
+          aiTracingContext,
+        };
+
         const telemetry = this.mastra?.getTelemetry();
         const span = executionContext.executionSpan;
         if (!telemetry || !span) {
-          return step.execute(data);
+          return step.execute(enhancedData);
         }
 
         return otlpContext.with(trace.setSpan(otlpContext.active(), span), async () => {
           return telemetry.traceMethod(step.execute.bind(step), {
             spanName,
             attributes,
-          })(data);
+          })(enhancedData);
         });
       };
     };
@@ -711,6 +818,14 @@ export class DefaultExecutionEngine extends ExecutionEngine {
               );
         this.logger.trackException(error);
         this.logger.error(`Error executing step ${step.id}: ` + error?.stack);
+
+        stepAISpan?.error({
+          error,
+          attributes: {
+            status: 'failed',
+          },
+        });
+
         execResults = {
           status: 'failed',
           error: error?.stack,
@@ -773,6 +888,15 @@ export class DefaultExecutionEngine extends ExecutionEngine {
           },
         });
       }
+    }
+
+    if (execResults.status != 'failed') {
+      stepAISpan?.end({
+        output: execResults.output,
+        attributes: {
+          status: execResults.status,
+        },
+      });
     }
 
     return { ...stepInfo, ...execResults };
@@ -1067,6 +1191,7 @@ export class DefaultExecutionEngine extends ExecutionEngine {
     abortController,
     runtimeContext,
     writableStream,
+    serializedStepGraph,
   }: {
     workflowId: string;
     runId: string;
@@ -1090,10 +1215,12 @@ export class DefaultExecutionEngine extends ExecutionEngine {
     abortController: AbortController;
     runtimeContext: RuntimeContext;
     writableStream?: WritableStream<ChunkType>;
+    serializedStepGraph: SerializedStepFlowEntry[];
   }): Promise<StepResult<any, any, any, any>> {
     const { step, condition } = entry;
     let isTrue = true;
-    let result = { status: 'success', output: prevOutput } as unknown as StepResult<any, any, any, any>;
+    const prevPayload = stepResults[step.id]?.payload;
+    let result = { status: 'success', output: prevPayload ?? prevOutput } as unknown as StepResult<any, any, any, any>;
     let currentResume = resume;
 
     do {
@@ -1109,6 +1236,7 @@ export class DefaultExecutionEngine extends ExecutionEngine {
         abortController,
         runtimeContext,
         writableStream,
+        serializedStepGraph,
       });
 
       // Clear resume for next iteration only if the step has completed resuming
@@ -1172,6 +1300,7 @@ export class DefaultExecutionEngine extends ExecutionEngine {
     abortController,
     runtimeContext,
     writableStream,
+    serializedStepGraph,
   }: {
     workflowId: string;
     runId: string;
@@ -1196,6 +1325,7 @@ export class DefaultExecutionEngine extends ExecutionEngine {
     abortController: AbortController;
     runtimeContext: RuntimeContext;
     writableStream?: WritableStream<ChunkType>;
+    serializedStepGraph: SerializedStepFlowEntry[];
   }): Promise<StepResult<any, any, any, any>> {
     const { step, opts } = entry;
     const results: StepResult<any, any, any, any>[] = [];
@@ -1259,6 +1389,7 @@ export class DefaultExecutionEngine extends ExecutionEngine {
             runtimeContext,
             skipEmits: true,
             writableStream,
+            serializedStepGraph,
           });
         }),
       );
@@ -1480,6 +1611,7 @@ export class DefaultExecutionEngine extends ExecutionEngine {
         abortController,
         runtimeContext,
         writableStream,
+        serializedStepGraph,
       });
     } else if (resume?.resumePath?.length && entry.type === 'parallel') {
       const idx = resume.resumePath.shift();
@@ -1626,6 +1758,7 @@ export class DefaultExecutionEngine extends ExecutionEngine {
         abortController,
         runtimeContext,
         writableStream,
+        serializedStepGraph,
       });
     } else if (entry.type === 'foreach') {
       execResults = await this.executeForeach({
@@ -1641,6 +1774,7 @@ export class DefaultExecutionEngine extends ExecutionEngine {
         abortController,
         runtimeContext,
         writableStream,
+        serializedStepGraph,
       });
     } else if (entry.type === 'sleep') {
       const startedAt = Date.now();
@@ -1933,16 +2067,6 @@ export class DefaultExecutionEngine extends ExecutionEngine {
       try {
         eventData = await this.executeWaitForEvent({ event: entry.event, emitter, timeout: entry.timeout });
 
-        await this.persistStepUpdate({
-          workflowId,
-          runId,
-          serializedStepGraph,
-          stepResults,
-          executionContext,
-          workflowStatus: 'running',
-          runtimeContext,
-        });
-
         const { step } = entry;
         execResults = await this.executeStep({
           workflowId,
@@ -1959,6 +2083,7 @@ export class DefaultExecutionEngine extends ExecutionEngine {
           abortController,
           runtimeContext,
           writableStream,
+          serializedStepGraph,
         });
       } catch (error) {
         execResults = {