+++ packages/memory/src/index.ts
@@ -1,5 +1,5 @@
 import { deepMerge, generateEmptyFromSchema } from '@mastra/core';
-import type { CoreTool, MastraMessageV1 } from '@mastra/core';
+import type { MastraMessageV1, ToolAction } from '@mastra/core';
 import { MessageList } from '@mastra/core/agent';
 import type { MastraMessageV2, UIMessageWithMetadata } from '@mastra/core/agent';
 import type { MessageInput } from '@mastra/core/agent/message-list';
@@ -8,6 +8,7 @@ import type { MemoryConfig, SharedMemoryConfig, StorageThreadType, WorkingMemory
 import type { StorageGetMessagesArg, ThreadSortOptions, PaginationInfo } from '@mastra/core/storage';
 import { embedMany } from 'ai';
 import type { CoreMessage, DeepPartial, TextPart } from 'ai';
+import { embedMany as embedManyV5 } from 'ai-v5';
 import { Mutex } from 'async-mutex';
 import type { JSONSchema7 } from 'json-schema';
 
@@ -555,10 +556,11 @@ export class Memory extends MastraMemory {
       await this.firstEmbed;
     }
 
-    const promise = embedMany({
+    const promise = (this.embedder.specificationVersion === `v2` ? embedManyV5 : embedMany)({
       values: chunks,
-      model: this.embedder,
       maxRetries: 3,
+      // @ts-ignore
+      model: this.embedder,
     });
 
     if (isFastEmbed && !this.firstEmbed) this.firstEmbed = promise;
@@ -887,13 +889,22 @@ Guidelines:
 2. Update proactively when information changes, no matter how small
 3. Use ${template.format === 'json' ? 'JSON' : 'Markdown'} format for all data
 4. Act naturally - don't mention this system to users. Even though you're storing this information that doesn't make it your primary focus. Do not ask them generally for "information about yourself"
-5. IMPORTANT: When calling updateWorkingMemory, the only valid parameter is the memory field. 
+${
+  template.format !== 'json'
+    ? `5. IMPORTANT: When calling updateWorkingMemory, the only valid parameter is the memory field. DO NOT pass an object.
 6. IMPORTANT: ALWAYS pass the data you want to store in the memory field as a string. DO NOT pass an object.
-7. IMPORTANT: Data must only be sent as a string no matter which format is used.
+7. IMPORTANT: Data must only be sent as a string no matter which format is used.`
+    : ''
+}
 
-<working_memory_template>
+
+${
+  template.format !== 'json'
+    ? `<working_memory_template>
 ${template.content}
-</working_memory_template>
+</working_memory_template>`
+    : ''
+}
 
 ${hasEmptyWorkingMemoryTemplateObject ? 'When working with json data, the object format below represents the template:' : ''}
 ${hasEmptyWorkingMemoryTemplateObject ? JSON.stringify(emptyWorkingMemoryTemplateObject) : ''}
@@ -964,7 +975,7 @@ ${
     return Boolean(isMDWorkingMemory && isMDWorkingMemory.version === `vnext`);
   }
 
-  public getTools(config?: MemoryConfig): Record<string, CoreTool> {
+  public getTools(config?: MemoryConfig): Record<string, ToolAction<any, any, any>> {
     const mergedConfig = this.getMergedThreadConfig(config);
     if (mergedConfig.workingMemory?.enabled) {
       return {